Use errors="replace" mode when decoding ffmpeg output
ffmpeg can output invalid UTF-8 from file tags, for example if the tag
is longer than 255 bytes, it might truncate the tag in the middle of a
UTF-8 sequence.

The exact ffmpeg output doesn't need to be round-tripped here, so use
the errors="replace" mode to stick a U+FFFD in there so processing can
continue.